pyproject.toml
[project]
name = "invoice-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.192.0",
    "google-auth>=2.49.1",
    "google-auth-oauthlib>=1.3.0",
    "google-genai>=1.67.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
]
